service.db to service.rrd.db
