+++ domain-rrd/service-rrd-db/src/test/java/gov/va/vro/service/rrd/db/TestConfig.java
@@ -1,0 +1,10 @@
+package gov.va.vro.service.rrd.db;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication(scanBasePackages = {"gov.va.vro.service.db"})
+@EnableJpaRepositories("gov.va.vro.persistence.repository")
+@EntityScan("gov.va.vro.persistence.model")
+public class TestConfig {}
