+++ domain-rrd/service-rrd-db/src/test/resources/init.sql
