+++ settings.gradle
@@ -81,7 +81,7 @@ include ':persistence:model'
 include ':scripts'
 include ':domain-rrd:services'
 include ':domain-rrd:spi'
-include ':domain-rrd:db'
+include ':domain-rrd:service-rrd-db'
 include ':domain-rrd:model-rrd'
 include ':shared:lib-camel-connector'
 include ':service-python'
